requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
rank_bm25
kiwipiepy
gradio
pandas
numpy
python-dotenv
matplotlib